HPPC-116: added more tests verifying empty key.

+++ hppc/src/test/templates/com/carrotsearch/hppc/AbstractKTypeTest.java
@@ -37,10 +37,10 @@ public abstract class AbstractKTypeTest<KType> extends RandomizedTest
     /**
      * Convert to target type from an integer used to test stuff. 
      */
-    public KType cast(int v)
+    public KType cast(Integer v)
     {
         /*! #if ($TemplateOptions.KTypePrimitive)
-            return (KType) v;
+            return (KType) v.intValue();
             #else !*/ 
             // @SuppressWarnings("unchecked")        
             KType k = (KType)(Object) v;
